+++ apps/cache-testing/tests/pages.spec.ts
@@ -1,5 +1,6 @@
 import Timers from 'node:timers/promises';
 import { expect, test } from '@playwright/test';
+import { revalidateByApi } from './test-helpers';
 
 const paths = [
     '/pages/with-paths/fallback-blocking/200',
@@ -326,5 +327,181 @@ test.describe('SSR', () => {
         const valueFromPageAfterReload = Number.parseInt((await page.getByTestId('data').innerText()).valueOf(), 10);
 
         expect(valueFromPageAfterReload - valueFromPage === 1).toBe(true);
+    });
+});
+
+test.describe('SSR neshClassicCache', () => {
+    const path = '/pages/no-paths/ssr-with-cache/200';
+
+    test('If revalidate is called, then page should be fresh after reload', async ({ page, baseURL }) => {
+        const url = new URL(path, `${baseURL}:3000`);
+
+        await page.goto(url.href);
+
+        const response = await revalidateByApi(path, `${baseURL}:3000`);
+
+        expect(response.revalidated).toBe(true);
+
+        await page.reload();
+
+        await expect(page.getByTestId('cache-state')).toContainText('fresh');
+    });
+
+    test('If revalidate is called on page A, then page B should be fresh on load and value must be updated', async ({
+        context,
+        baseURL,
+    }) => {
+        const appAUrl = new URL(path, `${baseURL}:3000`);
+
+        await revalidateByApi(path, `${baseURL}:3000`);
+
+        const appA = await context.newPage();
+
+        await appA.goto(appAUrl.href);
+
+        await expect(appA.getByTestId('cache-state')).toContainText('fresh');
+
+        const valueFromPageA = Number.parseInt((await appA.getByTestId('data').innerText()).valueOf(), 10);
+
+        await revalidateByApi(path, `${baseURL}:3000`);
+
+        const appBUrl = new URL(path, `${baseURL}:3001`);
+
+        const appB = await context.newPage();
+
+        await appB.goto(appBUrl.href);
+
+        await expect(appA.getByTestId('cache-state')).toContainText('fresh');
+
+        const valueFromPageB = Number.parseInt((await appB.getByTestId('data').innerText()).valueOf(), 10);
+
+        expect(valueFromPageB - valueFromPageA === 1).toBe(true);
+    });
+
+    test('Page should be fresh after becoming stale and reloaded twice', async ({ page, baseURL }) => {
+        const url = new URL(path, `${baseURL}:3000`);
+
+        await revalidateByApi(path, `${baseURL}:3000`);
+
+        await page.goto(url.href);
+
+        await expect(page.getByTestId('cache-state')).toContainText('fresh');
+
+        await expect(page.getByTestId('cache-state')).toContainText('stale', { timeout: 7500 });
+
+        await page.reload();
+
+        await expect(page.getByTestId('cache-state')).toContainText('stale');
+
+        await page.reload();
+
+        await expect(page.getByTestId('cache-state')).toContainText('fresh');
+    });
+});
+
+test.describe('API route neshClassicCache', () => {
+    const path = '/api/pages-cached-api';
+
+    test('If revalidate is called, then page should be fresh after reload', async ({ baseURL }) => {
+        const url = new URL(path, `${baseURL}:3000`);
+
+        const revalidateByApiResponse = await revalidateByApi(path, `${baseURL}:3000`);
+
+        expect(revalidateByApiResponse.revalidated).toBe(true);
+
+        const response = await fetch(url.href);
+
+        if (!response.ok) {
+            throw new Error('Failed to fetch');
+        }
+
+        const json = (await response.json()) as { props: { count: number; timeMs: string; revalidateAfterMs: string } };
+
+        const time = Number.parseInt(json.props.timeMs, 10);
+
+        const revalidateAfter = Number.parseInt(json.props.revalidateAfterMs, 10);
+
+        expect(time < revalidateAfter).toBe(true);
+    });
+
+    test('If revalidate is called on page A, then page B should be fresh on load and value must be updated', async ({
+        baseURL,
+    }) => {
+        const appAUrl = new URL(path, `${baseURL}:3000`);
+
+        const appBUrl = new URL(path, `${baseURL}:3001`);
+
+        await revalidateByApi(path, `${baseURL}:3000`);
+
+        const appAResponse = await fetch(appAUrl.href);
+
+        if (!appAResponse.ok) {
+            throw new Error('Failed to fetch');
+        }
+
+        const appAJson = (await appAResponse.json()) as {
+            props: { count: number; timeMs: string; revalidateAfterMs: string };
+        };
+
+        const time = Number.parseInt(appAJson.props.timeMs, 10);
+
+        const revalidateAfter = Number.parseInt(appAJson.props.revalidateAfterMs, 10);
+
+        expect(time < revalidateAfter).toBe(true);
+
+        const appACount = appAJson.props.count;
+
+        await revalidateByApi(path, `${baseURL}:3000`);
+
+        const appBResponse = await fetch(appBUrl.href);
+
+        if (!appBResponse.ok) {
+            throw new Error('Failed to fetch');
+        }
+
+        const appBJson = (await appBResponse.json()) as {
+            props: { count: number; timeMs: string; revalidateAfterMs: string };
+        };
+
+        const appBCount = appBJson.props.count;
+
+        expect(appBCount - appACount === 1).toBe(true);
+    });
+
+    test('Page should be fresh after becoming stale and reloaded twice', async ({ baseURL }) => {
+        const url = new URL(path, `${baseURL}:3000`);
+
+        await revalidateByApi(path, `${baseURL}:3000`);
+
+        const response1 = await fetch(url.href);
+
+        if (!response1.ok) {
+            throw new Error('Failed to fetch');
+        }
+
+        const json1 = (await response1.json()) as {
+            props: { count: number; timeMs: string; revalidateAfterMs: string };
+        };
+
+        const time = Number.parseInt(json1.props.timeMs, 10);
+
+        const revalidateAfter = Number.parseInt(json1.props.revalidateAfterMs, 10);
+
+        expect(time < revalidateAfter).toBe(true);
+
+        await Timers.setTimeout(revalidateAfter - Date.now());
+
+        await fetch(url.href);
+        const response2 = await fetch(url.href);
+
+        if (!response2.ok) {
+            throw new Error('Failed to fetch');
+        }
+
+        const json2 = (await response2.json()) as {
+            props: { count: number; timeMs: string; revalidateAfterMs: string };
+        };
+
+        expect(json2.props.count - json1.props.count === 1).toBe(true);
     });
 });