+++ docs/cache-handler-docs/src/pages/functions/_meta.ts
@@ -1,3 +1,4 @@
 export default {
     'nesh-cache': 'neshCache',
+    'nesh-classic-cache': 'neshClassicCache',
 };