+++ internal/next-lru-cache/src/cache-types/next-cache-handler-value.ts
@@ -1,4 +1,4 @@
-import type { CacheHandlerValue } from '@neshca/next-common';
+import type { CacheHandlerValue } from '@repo/next-common';
 import type { LRUCache } from 'lru-cache';
 
 import type { LruCacheOptions } from '../create-configured-cache';